move files; add dummy test
