+++ sentry-samples/sentry-samples-spring-boot-jakarta/build.gradle.kts
@@ -97,14 +97,9 @@ tasks.register<Test>("systemTest").configure {
 
 tasks.named("test").configure {
     require(this is Test)
-//    maxParallelForks = Runtime.getRuntime().availableProcessors() / 2
-//
-//    // Cap JVM args per test
-//    minHeapSize = "128m"
-//    maxHeapSize = "1g"
 
     filter {
-        excludeTestsMatching("io.sentry.samples.*")
+        excludeTestsMatching("io.sentry.systemtest.*")
     }
 }
