+++ sentry-samples/sentry-samples-spring-boot-jakarta/src/test/kotlin/io/sentry/DummyTest.kt
@@ -1,0 +1,13 @@
+package io.sentry
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class DummyTest {
+
+    @Test
+    fun `the only test`() {
+        // only needed to have more than 0 tests and not fail the build
+        assertTrue(true)
+    }
+}
