+++ sentry-samples/sentry-samples-spring-boot-jakarta/src/test/kotlin/io/sentry/systemtest/GraphqlProjectSystemTest.kt
@@ -1,6 +1,6 @@
-package io.sentry.samples.spring.boot.jakarta.systemtest
+package io.sentry.systemtest
 
-import io.sentry.samples.spring.boot.jakarta.systemtest.util.TestHelper
+import io.sentry.systemtest.util.TestHelper
 import org.junit.Before
 import kotlin.test.Test
 import kotlin.test.assertEquals
