+++ sentry-samples/sentry-samples-spring-boot-jakarta/src/test/kotlin/io/sentry/systemtest/TodoSystemTest.kt
@@ -1,6 +1,6 @@
-package io.sentry.samples.spring.boot.jakarta.systemtest
+package io.sentry.systemtest
 
-import io.sentry.samples.spring.boot.jakarta.systemtest.util.TestHelper
+import io.sentry.systemtest.util.TestHelper
 import org.junit.Before
 import org.springframework.http.HttpStatus
 import kotlin.test.Test
